Fix dbscript handling with 0 delay.
I wrongly assumed that i should take care only of first script in the list to handle db script with 0 delay.
There can be more than one with 0 as delay.
I also forget to take into accout the result of the script. If a step in the script fail we should stop the rest of the script.

Thank @Grz3s for pointing and helping.